added a cofirm alert when deleting a user

+++ frontend/src/screens/UserListScreen.js
@@ -31,7 +31,9 @@ const UserListScreen = () => {
   }, [dispatch, navigate, successDelete]);
 
   const deleteHandler = (id) => {
-    dispatch(deleteUser(id));
+    if (window.confirm("Are you sure you want to delete this user?")) {
+      dispatch(deleteUser(id));
+    }
   };
 
   return (
